Added confirmation before deletion of listings.

+++ app/Controllers/HousesController.js
@@ -2,6 +2,7 @@ import { appState } from "../AppState.js";
 import { House } from "../Models/House.js";
 import { housesService } from "../Services/HousesService.js";
 import { getFormData } from "../Utils/FormHandler.js";
+import { Pop } from "../Utils/Pop.js";
 import { setHTML } from "../Utils/Writer.js";
 
 function _drawHouses() {
@@ -40,8 +41,9 @@ export class HousesController {
     form.reset()
   }
 
-  deleteHouse(id) {
-    housesService.deleteHouse(id)
+  async deleteHouse(id) {
+    if (await Pop.confirm("Are you sure you want to delete this listing?"))
+      housesService.deleteHouse(id)
   }
 
 }
